refactored to add idempotence/abstraction and remove duplication
Joined separate 'update template for category and pkg, then POST XML'
calls into one customizeAndPostXMLtoAPI. Made pkg copy idempotent.
TODO: all previous and output variables

+++ Code/autopkglib/JSSImporter.py
@@ -23,7 +23,7 @@ import shutil
 from distutils import version
 
 from autopkglib import Processor, ProcessorError
-from autopkglib import Copier
+from autopkglib import Copier # not successfully leveraging yet
 
 __all__ = ["JSSImporter"]
 
@@ -72,14 +72,14 @@ class JSSImporter(Processor):
             "description": "Name of static group to create with which to offer item to users with an older version.",
         },
     }
-    output_variables = {}
-    #     "jss_category_added": {
-    #         "description": "True if category was created."
-    #     },
-    #     "jss_repo_changed": {
-    #         "description": "True if item was imported."
-    #     },
-    # }
+    output_variables = {
+        # "jss_category_added": {
+        #     "description": "True if category was created."
+        # },
+        # "jss_repo_changed": {
+        #     "description": "True if item was imported."
+        # },
+    }
     description = __doc__
 
     def checkItem(self, repoUrl, base64string, item_to_check, apiUrl):
@@ -101,9 +101,10 @@ class JSSImporter(Processor):
             raise RuntimeError('Did not get a valid response from the server')
         # assign fetched JSS categories to a string and search if prod name already exists
         jss_results = submitResult.read()
-        category_match = re.search(re.escape(item_to_check), jss_results)
-        if category_match:
+        item_match = re.search(re.escape(item_to_check), jss_results)
+        if item_match:
             return "proceed"
+            self.output("Found %s, moving on" % item_to_check)
         else:
             bracket = "<id>"
             nums = re.findall(re.escape(bracket) + r'\d+', jss_results)
@@ -112,27 +113,15 @@ class JSSImporter(Processor):
             highest_id += 1
             return highest_id
 
-    def customizeAndPostPackage(self, repoUrl, pkg_id, pkg_name, prod_name, template_string, base64string):
+    def customizeAndPostXMLtoAPI(self, repoUrl, apiUrl, item_id, replace_dict, template_string, base64string):
         """After finding an unused id, this updates a template with the id and product name for a category"""
-        replace_dict = {"%PKG_ID%" : pkg_id, "%PKG_NAME%" : pkg_name, "%PROD_NAME%" : prod_name}
         for key, value in replace_dict.iteritems():
             template_string = template_string.replace(key, value)
-        submitRequest = urllib2.Request(repoUrl + "/JSSResource/packages/id/" + pkg_id, template_string, {'Content-type': 'text/xml'})
+        submitRequest = urllib2.Request(repoUrl + "/JSSResource/" + apiUrl + "/id/" + item_id, template_string, {'Content-type': 'text/xml'})
         submitRequest.add_header("Authorization", "Basic %s" % base64string)
         submitResult = urllib2.urlopen(submitRequest)
         submittedResult = submitResult.read()
-        self.output("Added package to JSS via API")
-
-    def customizeAndPostCategory(self, repoUrl, cat_id, cat_name, template_string, base64string):
-        """After finding an unused id, this updates a template with the id and product name for a category"""
-        replace_dict = {"%CAT_ID%" : cat_id, "%CAT_NAME%" : cat_name}
-        for key, value in replace_dict.iteritems():
-            template_string = template_string.replace(key, value)
-        submitRequest = urllib2.Request(repoUrl + "/JSSResource/categories/id/" + cat_id, template_string, {'Content-type': 'text/xml'})
-        submitRequest.add_header("Authorization", "Basic %s" % base64string)
-        submitResult = urllib2.urlopen(submitRequest)
-        submittedResult = submitResult.read()
-        self.output("Added category to JSS via API")
+        self.output("Added to %s section of JSS via API" % apiUrl)
 
     def main(self):
         repoUrl = self.env["jss_url"]
@@ -151,32 +140,38 @@ class JSSImporter(Processor):
             template_string = """<?xml version="1.0" encoding="UTF-8"?><category><id>%CAT_ID%</id><name>%CAT_NAME%</name></category>"""
             if highest_id != "proceed":
                 highest_id = str(highest_id)
-                self.customizeAndPostCategory(repoUrl, highest_id, prod_name, template_string, base64string)
+                replace_dict = {"%CAT_ID%" : highest_id, "%CAT_NAME%" : prod_name}                
+                self.customizeAndPostXMLtoAPI(repoUrl, apiUrl, highest_id, replace_dict, template_string, base64string)
+                # self.env["jss_category_added"] = True
+        # else:
+        #     if not "jss_category_added" in self.env:
+        #         self.env["jss_category_added"] = False
         # start metadata/pkg upload
         template_string = """<?xml version="1.0" encoding="UTF-8"?><package><id>%PKG_ID%</id><name>%PKG_NAME%</name><category>%PROD_NAME%</category><filename>%PKG_NAME%</filename><info/><notes/><priority>10</priority><reboot_required>false</reboot_required><fill_user_template>false</fill_user_template><fill_existing_users>false</fill_existing_users><boot_volume_required>false</boot_volume_required><allow_uninstalled>false</allow_uninstalled><os_requirements/><required_processor>None</required_processor><switch_with_package>Do Not Install</switch_with_package><install_if_reported_available>false</install_if_reported_available><reinstall_option>Do Not Reinstall</reinstall_option><triggering_files/><send_notification>false</send_notification></package>"""
         apiUrl = "packages"
         item_to_check = "<name>" + pkg_name + "</name>"
         highest_id = self.checkItem(repoUrl, base64string, item_to_check, apiUrl)
-        source_item = self.env["pkg_path"]
-        dest_item = (self.env["repo_path"] + "/" + pkg_name)
-        try:
-            shutil.copyfile(source_item, dest_item)
-            self.output("Copied %s to %s" % (source_item, dest_item))
-        except BaseException, err:
-            raise ProcessorError(
-                "Can't copy %s to %s: %s" % (source_item, dest_item, err))        
-        apiUrl = "packages"
-        #run me some main, why dontcha, capturing the checkCategory stages output
-        highest_id = self.checkItem(repoUrl, base64string, item_to_check, apiUrl)
         # if prod name already exists then we'd proceed to the next processor,
         if highest_id != "proceed":
             pkg_id = str(highest_id)
-            self.customizeAndPostPackage(repoUrl, pkg_id, pkg_name, prod_name, template_string, base64string)        
-        
-        # set output variables
-        self.env["jss_repo_changed"] = True
-        
-        self.output("Copied pkg to %s" % self.env["repo_path"])
+            replace_dict = {"%PKG_ID%" : pkg_id, "%PKG_NAME%" : pkg_name, "%PROD_NAME%" : prod_name}
+            self.customizeAndPostXMLtoAPI(repoUrl, apiUrl, pkg_id, replace_dict, template_string, base64string)
+        source_item = self.env["pkg_path"]
+        dest_item = (self.env["repo_path"] + "/" + pkg_name)
+        if os.path.exists(dest_item):
+            self.output("Pkg already exists at %s, moving on" % dest_item)
+        else:
+            try:
+                shutil.copyfile(source_item, dest_item)
+                self.output("Copied %s to %s" % (source_item, dest_item))
+            except BaseException, err:
+                raise ProcessorError(
+                    "Can't copy %s to %s: %s" % (source_item, dest_item, err))
+            # set output variables
+        #     self.env["jss_repo_changed"] = True
+        # else:
+        #     if not "jss_repo_changed" in self.env:
+        #         self.env["jss_repo_changed"] = False
 
 if __name__ == "__main__":
     processor = JSSImporter()
